session-persistence: Rename files to follow the prefix-name rule
The files implement functions in the XdpSessionPersistence namespace so
rename the files accordingly.
