+++ src/xdp-session-persistence.c
@@ -22,7 +22,7 @@
 #include "config.h"
 
 #include "permissions.h"
-#include "restore-token.h"
+#include "xdp-session-persistence.h"
 
 static GMutex transient_permissions_lock;
 static GHashTable *transient_permissions;
